feat(as-4929): swap urls for new-plans-drawings and plans-drawings
