+++ packages/forms-web-app/src/views/full-appeal/submit-appeal/plans-drawings.njk
@@ -1,16 +1,18 @@
 {% extends "layouts/full-appeal-banner/main.njk" %}
 
 {% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
-{% from "govuk/components/radios/macro.njk" import govukRadios %}
+{% from "govuk/components/file-upload/macro.njk" import govukFileUpload %}
+{% from "govuk/components/details/macro.njk" import govukDetails %}
 {% from "govuk/components/button/macro.njk" import govukButton %}
 
-{% set title = "Plans and drawings - Appeal a planning decision - GOV.UK" %}
+{% set labelText = "Drag and drop or choose files" %}
+{% set title = "Plans or drawings - Appeal a planning decision - GOV.UK" %}
 {% block pageTitle %}{{ "Error: " + title if errors else title }}{% endblock %}
 
 {% block backButton %}
   {{ govukBackLink({
     text: 'Back',
-    href: '/full-appeal/submit-appeal/appeal-statement',
+    href: '/full-appeal/submit-appeal/new-plans-drawings',
     attributes: {
       'data-cy': 'back'
     }
@@ -27,37 +29,49 @@
           attributes: {"data-cy": "error-wrapper"}
         }) }}
       {% endif %}
-      <form action="" method="post" novalidate>
-        <span class="govuk-caption-l">Upload documents for your appeal</span>
-        {{ govukRadios({
-          classes: "govuk-radios",
-          idPrefix: "plans-drawings",
-          name: "plans-drawings",
-          errorMessage: errors['plans-drawings'] and {
-            text: errors['plans-drawings'].msg
+      <span class="govuk-caption-l">Upload documents for your appeal</span>
+      <h1 class="govuk-heading-l">Plans or drawings</h1>
+      {% if uploadedFiles.length %}
+        <h2 class="govuk-heading-m govuk-!-margin-top-8">Uploaded files</h2>
+        <ul class="govuk-list">
+        {%- for file in uploadedFiles %}
+          <li><a href="/document/{{appealId}}/{{file.id}}" class="govuk-link">{{ file.name }}</a></li>
+        {%- endfor %}
+        </ul>
+        {% set labelText = "Replace the files" %}
+      {% endif %}
+      <form action="" method="post" novalidate enctype="multipart/form-data">
+        {{ govukFileUpload({
+          id: "file-upload",
+          name: "file-upload",
+          classes: "pins-file-upload",
+          attributes: {
+            'multiple': true
           },
-          fieldset: {
-            legend: {
-              text: "Do you have any new plans or drawings that support your appeal?",
-              isPageHeading: true,
-              classes: "govuk-fieldset__legend--l"
-            }
+          label: {
+            text: labelText,
+            classes: "govuk-label--m"
           },
-          items: [
-            {
-              value: "yes",
-              text: "Yes",
-              attributes: { "data-cy": "answer-yes" },
-              checked: hasPlansDrawings == true
-            },
-            {
-              value: "no",
-              text: "No",
-              attributes: { "data-cy": "answer-no" },
-              checked: hasPlansDrawings == false
-            }
-          ]
+          errorMessage: errors['file-upload'] and { text: errors['file-upload'].msg }
         }) }}
+
+        {{ govukDetails({
+          summaryText: "Files you can upload",
+          html: '<p class="govuk-body">
+                  You can upload the following types of file:
+                </p>
+                <ul class="govuk-list govuk-list--bullet">
+                  <li>DOC or DOCX</li>
+                  <li>JPG or JPEG</li>
+                  <li>PDF</li>
+                  <li>PNG</li>
+                  <li>TIF or TIFF</li>
+                </ul>
+                <p class="govuk-body">
+                  Your file must be smaller than 15MB.
+                </p>'
+        }) }}
+
         {{ govukButton({
           text: "Continue",
           type: "submit",
@@ -66,5 +80,4 @@
       </form>
     </div>
   </div>
-
 {% endblock %}
